Reorganized tests update build script to execute all integration tests
