+++ test/integration/aws/cloudformation_validation.py
